+++ readium/streamer/r2-streamer/src/main/java/org/readium/r2/streamer/parser/readium/ReadiumWebPubParser.kt
@@ -9,15 +9,10 @@
 
 package org.readium.r2.streamer.parser.readium
 
-import android.content.Context
 import kotlinx.coroutines.runBlocking
-import org.json.JSONObject
 import org.readium.r2.shared.PdfSupport
 import org.readium.r2.shared.drm.DRM
-import org.readium.r2.shared.fetcher.ArchiveFetcher
-import org.readium.r2.shared.fetcher.Fetcher
-import org.readium.r2.shared.fetcher.FileFetcher
-import org.readium.r2.shared.fetcher.TransformingFetcher
+import org.readium.r2.shared.fetcher.*
 import org.readium.r2.shared.publication.Manifest
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.shared.publication.asset.FileAsset
@@ -25,13 +20,14 @@ import org.readium.r2.shared.publication.asset.PublicationAsset
 import org.readium.r2.shared.publication.services.PerResourcePositionsService
 import org.readium.r2.shared.publication.services.locatorServiceFactory
 import org.readium.r2.shared.publication.services.positionsServiceFactory
+import org.readium.r2.shared.util.http.HttpClient
 import org.readium.r2.shared.util.logging.WarningLogger
 import org.readium.r2.shared.util.mediatype.MediaType
 import org.readium.r2.shared.util.pdf.PdfDocumentFactory
-import org.readium.r2.streamer.DefaultPdfDocumentFactory
 import org.readium.r2.streamer.PublicationParser
 import org.readium.r2.streamer.container.ContainerError
 import org.readium.r2.streamer.container.PublicationContainer
+import org.readium.r2.streamer.extensions.readAsJsonOrNull
 import org.readium.r2.streamer.fetcher.LcpDecryptor
 import org.readium.r2.streamer.parser.PubBox
 import org.readium.r2.streamer.parser.audio.AudioLocatorService
@@ -43,38 +39,46 @@ import java.io.FileNotFoundException
  * Parses any Readium Web Publication package or manifest, e.g. WebPub, Audiobook, DiViNa, LCPDF...
  */
 @OptIn(PdfSupport::class)
-class ReadiumWebPubParser(private val pdfFactory: PdfDocumentFactory? = null) : PublicationParser, org.readium.r2.streamer.parser.PublicationParser {
-
-    constructor(context: Context) : this(pdfFactory = DefaultPdfDocumentFactory(context))
+class ReadiumWebPubParser(
+    private val pdfFactory: PdfDocumentFactory?,
+    private val httpClient: HttpClient,
+) : PublicationParser, org.readium.r2.streamer.parser.PublicationParser {
 
     override suspend fun parse(
         asset: PublicationAsset,
         fetcher: Fetcher,
         warnings: WarningLogger?
     ): Publication.Builder? {
+        val mediaType = asset.mediaType()
 
-        if (!asset.mediaType().isReadiumWebPubProfile)
+        if (!mediaType.isReadiumWebPubProfile)
             return null
 
-        val manifest =
-            if (asset.mediaType().isRwpm) {
-                val manifestLink = fetcher.links().firstOrNull()
-                    ?: error("Empty fetcher.")
-                val manifestJson = fetcher.get(manifestLink).use {
-                    it.readAsString().getOrThrow()
-                }
-                Manifest.fromJSON(JSONObject(manifestJson))
-            } else {
-                val manifestLink = fetcher.links()
-                    .firstOrNull { it.href == "/manifest.json" }
-                    ?: error("Unable to find a manifest link.")
-                val manifestJson = fetcher.get(manifestLink).use {
-                    it.readAsString().getOrThrow()
-                }
-                Manifest.fromJSON(JSONObject(manifestJson), packaged = true)
-            }
-                ?: throw Exception("Failed to parse RWPM.")
+        val isPackage = !mediaType.isRwpm
 
+        val manifestJson =
+            if (isPackage) {
+                fetcher.readAsJsonOrNull("/manifest.json")
+            } else {
+                // For a single manifest file, reads the first (and only) file in the fetcher.
+                fetcher.links().firstOrNull()
+                    ?.let { fetcher.readAsJsonOrNull(it.href) }
+            }
+                ?: throw Exception("Manifest not found")
+
+        val manifest = Manifest.fromJSON(manifestJson, packaged = isPackage)
+            ?: throw Exception("Failed to parse the RWPM Manifest")
+
+        @Suppress("NAME_SHADOWING")
+        var fetcher = fetcher
+
+        // For a manifest, we discard the [fetcher] provided by the Streamer, because it was only
+        // used to read the manifest file. We use an [HttpFetcher] instead to serve the remote
+        // resources.
+        if (!isPackage) {
+            val baseUrl = manifest.linkWithRel("self")?.let { File(it.href).parent }
+            fetcher = HttpFetcher(httpClient, baseUrl)
+        }
 
         // Checks the requirements from the LCPDF specification.
         // https://readium.org/lcp-specs/notes/lcp-for-pdf.html