+++ readium/streamer/r2-streamer/src/test/java/org/readium/r2/streamer/parser/epub/EpubPositionsServiceTest.kt
@@ -426,7 +426,7 @@ class EpubPositionsServiceTest {
 
                 override suspend fun length() = findResource(link.href)
                     ?.let { Try.success(it.first) }
-                    ?: Try.failure(Resource.Exception.NotFound)
+                    ?: Try.failure(Resource.Exception.NotFound())
 
                 override suspend fun read(range: LongRange?): ResourceTry<ByteArray> = Try.success(ByteArray(0))